Revert "Changed salary data type from int to double"
This reverts commit 7c1253ecfc4d7331972868cecddb4555cdfaa981.

+++ src/main/java/org/gitsample/Employee.java
@@ -1,10 +1,11 @@
 package org.gitsample;
 
 public class Employee {
-	// New comment added
+	
+	// Sample change 123
 	private String name;
 	private int age;
-	private double salary;
+	private int salary;
 	private String address;
 	
 	public String getName() {
